+++ src/pieces/bishop.hpp
@@ -6,9 +6,9 @@
 
 #include "utility/binary.hpp"
 
-std::uint64_t get_bishop_xrayed_squares_from_mailbox(std::uint8_t mb) noexcept;
-std::uint64_t get_bishop_blocker_squares_from_mailbox(std::uint8_t mb) noexcept;
-std::uint64_t get_bishop_attacked_squares_from_mailbox(std::uint64_t pos, std::uint8_t mb) noexcept;
+std::uint64_t get_bishop_xrayed_squares_from_mailbox(std::size_t mb) noexcept;
+std::uint64_t get_bishop_blocker_squares_from_mailbox(std::size_t mb) noexcept;
+std::uint64_t get_bishop_attacked_squares_from_mailbox(std::uint64_t pos, std::size_t mb) noexcept;
 
 inline std::uint64_t get_bishop_attacked_squares_from_mailboxes(std::uint64_t pos, auto... mbs) noexcept { return (get_bishop_attacked_squares_from_mailbox(pos, mbs) | ...); }
 
@@ -21,12 +21,12 @@ inline std::uint64_t get_bishop_attacked_squares_from_mailboxes(std::uint64_t po
 namespace details
 {
 
-inline std::uint64_t get_bishop_xrayed_squares_from_mailbox_impl(std::uint8_t mb) noexcept
+inline std::uint64_t get_bishop_xrayed_squares_from_mailbox_impl(std::size_t mb) noexcept
 {
     const std::uint64_t b { get_bitboard_mailbox_piece(mb) };
 
     std::uint64_t sw_ray {b}, nw_ray {b}, ne_ray {b}, se_ray {b};
-    for (std::uint8_t i = 0; i < 7; i++)
+    for (std::size_t i = 0; i < 7; i++)
     {
         sw_ray = get_bitboard_mailbox_piece(mb) | shift_south_west(sw_ray);
         nw_ray = get_bitboard_mailbox_piece(mb) | shift_north_west(nw_ray);
@@ -39,14 +39,14 @@ inline std::uint64_t get_bishop_xrayed_squares_from_mailbox_impl(std::uint8_t mb
 
 extern const std::span<std::uint64_t> xray_table_bishop;
 
-inline std::uint64_t get_bishop_blocker_squares_from_mailbox_impl(std::uint8_t mb) noexcept
+inline std::uint64_t get_bishop_blocker_squares_from_mailbox_impl(std::size_t mb) noexcept
 {
     return get_bishop_xrayed_squares_from_mailbox_impl(mb) & ~(FILE_A | FILE_H | RANK_1 | RANK_8);
 }
 
 extern const std::span<std::uint64_t> blocker_table_bishop;
 
-inline std::uint64_t get_bishop_attacked_squares_from_mailbox_impl(std::uint8_t mb, std::uint64_t pos)
+inline std::uint64_t get_bishop_attacked_squares_from_mailbox_impl(std::size_t mb, std::uint64_t pos)
 {
     std::uint64_t ret {};
 
@@ -71,17 +71,17 @@ extern const std::array<details::pext_bitboard, 64>  attack_table_bishop;
 
 }
 
-inline std::uint64_t get_bishop_xrayed_squares_from_mailbox(std::uint8_t mb) noexcept
+inline std::uint64_t get_bishop_xrayed_squares_from_mailbox(std::size_t mb) noexcept
 {
     return details::xray_table_bishop[mb];
 }
 
-inline std::uint64_t get_bishop_blocker_squares_from_mailbox(std::uint8_t mb) noexcept
+inline std::uint64_t get_bishop_blocker_squares_from_mailbox(std::size_t mb) noexcept
 {
     return details::blocker_table_bishop[mb];
 }
 
-inline std::uint64_t get_bishop_attacked_squares_from_mailbox(std::uint64_t pos, std::uint8_t mb) noexcept
+inline std::uint64_t get_bishop_attacked_squares_from_mailbox(std::uint64_t pos, std::size_t mb) noexcept
 {
     return details::attack_table_bishop[mb][pos];
 }
